Fix drop table migration

+++ db/migrate/20161201161337_remove_admin_users.rb
@@ -1,5 +1,5 @@
 class RemoveAdminUsers < ActiveRecord::Migration
   def change
-    drop_table :admin_users
+    drop_table :admin_users if ActiveRecord::Base.connection.table_exists? 'admin_users'
   end
 end
